requirements.txt
django==5.0
django-active-link==0.2.2
django-adminlte3==0.1.6
django-ajax-datatable==4.5.0
django-auto-logout==0.5.1
django-bootstrap-modal-forms==3.0.5
django-chartjs==2.3.0
django-google-fonts==0.0.3
django-simple-captcha==0.6.2
openpyxl==3.1.5
psycopg2==2.9.10
pyexcel==0.7.3
xhtml2pdf==0.2.11
weasyprint==66.0
python-dotenv==1.1.1
django-widget-tweaks==1.5.1